Other entities operations requirement handlers

+++ Services/RepairService.cs
@@ -1,7 +1,10 @@
 ﻿using AutoMapper;
+using CarAPI.Authorization;
 using CarAPI.Entities;
+using CarAPI.Enums;
 using CarAPI.Exceptions;
 using CarAPI.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 
 namespace CarAPI.Services
@@ -11,11 +14,15 @@ namespace CarAPI.Services
         private readonly CarDbContext _context;
         private readonly IMapper _mapper;
         private readonly ILogger _logger;
-        public RepairService(CarDbContext context, IMapper mapper, ILogger<TechnicalReviewService> logger)
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
+        public RepairService(CarDbContext context, IMapper mapper, ILogger<TechnicalReviewService> logger, IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _context = context;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
         private Car GetCarById(int carId) 
         {
@@ -33,10 +40,16 @@ namespace CarAPI.Services
 
             var repairEntity = _mapper.Map<Repair>(dto);
             repairEntity.CarId = carId;
+            repairEntity.AddedByUserId = _userContextService.UserId;
+            var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, repairEntity, new ResourceOperationRequirement(ResourceOperationType.Update)).Result;
+            if (!authResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
 
             _context.Repairs.Add(repairEntity);
             _context.SaveChanges();
-            _logger.LogInformation($"Technical review with id: {repairEntity.Id} has been created (car id: {carId})");
+            _logger.LogInformation($"Repair with id: {repairEntity.Id} has been created (car id: {carId}) by user with id: {_userContextService.UserId}");
 
             return repairEntity.Id;
         }
@@ -68,21 +81,32 @@ namespace CarAPI.Services
 
         public void DeleteAll(int carId)
         {
-            _logger.LogWarning($"Repairs delete action invoked (car id: {carId})");
+            _logger.LogWarning($"Repairs delete action invoked (car id: {carId}) by user with id: {_userContextService.UserId}");
             var car = GetCarById(carId);
+            var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, car, new ResourceOperationRequirement(ResourceOperationType.Delete)).Result;
+            if (!authResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
             _context.RemoveRange(car.CarRepairs);
             _context.SaveChanges();
+            _logger.LogWarning($"Repairs deleted (car id: {carId}) by user with id: {_userContextService.UserId}");
         }
 
         public void DeleteById(int carId, int repairId)
         {
-            _logger.LogWarning($"Repair with id: {repairId} delete action invoked (car id: {carId})");
+            _logger.LogWarning($"Repair with id: {repairId} delete action invoked (car id: {carId}) by user with id: {_userContextService.UserId}");
             var car = GetCarById(carId);
             var repair = GetRepairById(repairId);
             if (repair.CarId != carId) throw new ContentNotFoundException($"Provided car id is wrong (car id: {carId})");
-
+            var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, repair, new ResourceOperationRequirement(ResourceOperationType.Delete)).Result;
+            if (!authResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
             _context.Repairs.Remove(repair);
             _context.SaveChanges();
+            _logger.LogWarning($"Repair deleted (repair id: {repair.Id}) by user with id: {_userContextService.UserId}");
         }
 
         public void UpdateRepair(int carId, int repairId, UpdateRepairDto dto)
@@ -90,12 +114,17 @@ namespace CarAPI.Services
             var car = GetCarById(carId);
             var repair = GetRepairById(repairId);
             if (repair.CarId != carId) throw new ContentNotFoundException($"Provided car id is wrong (car id: {carId})");
+            var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, repair, new ResourceOperationRequirement(ResourceOperationType.Update)).Result;
+            if (!authResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
             _mapper.Map(dto, repair);
 
             repair.CarId = carId;
 
             _context.SaveChanges();
-            _logger.LogWarning($"Repair with id: {repairId}  has been updated (car id: {carId})");
+            _logger.LogWarning($"Repair with id: {repairId}  has been updated (car id: {carId}) by user with id:  {_userContextService.UserId}");
         }
     }
 }
