+++ Services/TechnicalReviewService.cs
@@ -1,7 +1,10 @@
 ﻿using AutoMapper;
+using CarAPI.Authorization;
 using CarAPI.Entities;
+using CarAPI.Enums;
 using CarAPI.Exceptions;
 using CarAPI.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 
 namespace CarAPI.Services
@@ -11,6 +14,8 @@ namespace CarAPI.Services
         private readonly CarDbContext _context;
         private readonly IMapper _mapper;
         private readonly ILogger _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
         public TechnicalReviewService(CarDbContext context, IMapper mapper, ILogger<TechnicalReviewService> logger)
         {
             _context = context;
@@ -22,12 +27,17 @@ namespace CarAPI.Services
             var car = GetCarById(carId);         
 
             var technicalReviewEntity = _mapper.Map<TechnicalReview>(dto);
-
+            technicalReviewEntity.AddedByUserId = _userContextService.UserId;
+            var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, technicalReviewEntity, new ResourceOperationRequirement(ResourceOperationType.Update)).Result;
+            if (!authResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
             technicalReviewEntity.CarId = carId;
 
             _context.TechnicalReviews.Add(technicalReviewEntity);
             _context.SaveChanges();
-            _logger.LogInformation($"Technical review with id: {technicalReviewEntity.Id} has been created (car id: {carId})");
+            _logger.LogInformation($"Technical review with id: {technicalReviewEntity.Id} has been created (car id: {carId}) by user with an Id: {_userContextService.UserId}");
 
             return technicalReviewEntity.Id;
         }
@@ -54,11 +64,16 @@ namespace CarAPI.Services
 
         public void DeleteAll(int carId)
         {
-            _logger.LogWarning($"Technical reviews delete action invoked (car id: {carId})");
+            _logger.LogWarning($"Technical reviews delete action invoked (car id: {carId}) by user with an Id: {_userContextService.UserId}");
             var car = GetCarById(carId);
+            var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, car, new ResourceOperationRequirement(ResourceOperationType.Delete)).Result;
+            if (!authResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }      
             _context.RemoveRange(car.TechnicalReviews);
             _context.SaveChanges();
-          
+            _logger.LogWarning($"Technical reviews deleted (car id: {carId}) by user with an Id: {_userContextService.UserId}");
         }
         private Car GetCarById(int carId)
         {
@@ -78,13 +93,19 @@ namespace CarAPI.Services
         }
         public void DeleteById(int carId, int technicalReviewId)
         {
-            _logger.LogWarning($"Technical review with id: {technicalReviewId} delete action invoked (car id: {carId})");
+            _logger.LogWarning($"Technical review with id: {technicalReviewId} delete action invoked (car id: {carId}) by user with an Id: {_userContextService.UserId}");
             var car = GetCarById(carId);
             var technicalReview = GetTechnicalReviewById(technicalReviewId);
+            var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, technicalReview, new ResourceOperationRequirement(ResourceOperationType.Delete)).Result;
+            if (!authResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
             if (technicalReview.CarId != carId) throw new ContentNotFoundException($"Provided car id is wrong (car id: {carId})");
 
             _context.TechnicalReviews.Remove(technicalReview);
             _context.SaveChanges();
+            _logger.LogWarning($"Technical review deleted (technical review id: {technicalReview.Id}) by user with an Id: {_userContextService.UserId}");
         }
 
         public void UpdateTechnicalReview(int carId, int technicalReviewId, UpdateTechnicalReviewDto dto)
@@ -92,12 +113,17 @@ namespace CarAPI.Services
             var car = GetCarById(carId);
             var technicalReview = GetTechnicalReviewById(technicalReviewId);
             if (technicalReview.CarId != carId) throw new ContentNotFoundException($"Provided car id is wrong (car id: {carId})");
+            var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, technicalReview, new ResourceOperationRequirement(ResourceOperationType.Update)).Result;
+            if (!authResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
             _mapper.Map(dto, technicalReview);
 
             technicalReview.CarId = carId;
 
             _context.SaveChanges();
-            _logger.LogWarning($"Technical review with id: {technicalReviewId}  has been updated (car id: {carId})");
+            _logger.LogWarning($"Technical review with id: {technicalReviewId}  has been updated (car id: {carId}) by user with an Id:  {_userContextService.UserId}");
 
 
         }
